day13/test_solver.py
import pytest
from solver import parse, solve1, solve2

TESTDATA = """
Button A: X+94, Y+34
Button B: X+22, Y+67
Prize: X=8400, Y=5400

Button A: X+26, Y+66
Button B: X+67, Y+21
Prize: X=12748, Y=12176

Button A: X+17, Y+86
Button B: X+84, Y+37
Prize: X=7870, Y=6450

Button A: X+69, Y+23
Button B: X+27, Y+71
Prize: X=18641, Y=10279
"""

@pytest.fixture
def parsed_data():
    return parse(TESTDATA)


def test_parse():
    data = parse(TESTDATA)
    assert data[0] == ((94, 34), (22, 67), (8400, 5400))


# PART 1
def test_solve1(parsed_data):
    solution = solve1(parsed_data)
    assert solution == 480


# PART 2
def test_solve2(parsed_data):
    solution = solve2(parsed_data)
    # asserts go here
